Changed test module filename to align with others
